requirements.txt
python-benedict==0.18.0
geopy==1.21.0
beautifulsoup4==4.8.2